+++ lib/models/genn_gene.dart
@@ -1,6 +1,7 @@
 part of 'package:genetically_evolving_neural_network/genetically_evolving_neural_network.dart';
 
 /// An extension of [GeneticEvolution.Gene].
+@JsonSerializable()
 class GENNGene extends Gene<GENNPerceptron> {
   const GENNGene({
     required super.value,
@@ -24,4 +25,12 @@ class GENNGene extends Gene<GENNPerceptron> {
       mutatedWaves: mutatedWaves ?? this.mutatedWaves,
     );
   }
+
+  /// Converts the input [json] into a [GENNGene] object.
+  factory GENNGene.fromJson(Map<String, dynamic> json) =>
+      _$GENNGeneFromJson(json);
+
+  /// Converts the [GENNGene] object to JSON.
+  @override
+  Map<String, dynamic> toJson() => _$GENNGeneToJson(this);
 }